--- linode_api/objects/support/__init__.py
@@ -1,2 +1,0 @@
-from .ticket import SupportTicket
-from .reply import TicketReply
